Adding content about header alignment
As per a question that came up, I added a sentence describing that column headers should align with the content (aka if your content is center aligned, center align the header as well). 
Open to suggestions of where it could go - I didn't include a lot of detail because there was already a lot going on in that list item.