fix(docs): updated width on table of contents (#4202)
It looks like some of the widths changed since v5. This caused flex to stop wrapping the table of contents at the viewports mentioned in https://github.com/patternfly/patternfly-org/issues/4178. The table of contents was still present, but just at the bottom of the screen since the order was 1 (the main content is ordered at 0). Dropping down to this width seems to allow wrapping as expected again.

Should fix https://github.com/patternfly/patternfly-org/issues/4178